requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
idna==2.7
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Money-Rocket==1.0.0
more-itertools==4.3.0
numpy==1.15.4
pandas==0.23.4
patsy==0.5.1
python-dateutil==2.7.5
pytz==2018.7
Quandl==3.4.4
requests==2.20.1
six==1.11.0
statsmodels==0.9.0
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
